requirements.txt
pytest==8.3.2
httpx==0.27.2
pydantic==2.9.0
pydantic-settings==2.6.0
allure-pytest==2.13.5
python-dotenv==1.0.1
Faker==28.4.1
